pyproject.toml
[project]
name = "nonebot-plugin-bf1tools"
version = "0.0.4"
description = "A plugin that interacts with the project [BF1ClientAPI]"
authors = ["SAFEluren"]
readme = "README.md"
requires-python = ">=3.11, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["nonebot_plugin_bf1tools"]
builtin_plugins = []

[tool.poetry]
name = "nonebot-plugin-bf1tools"
version = "0.0.1"
description = "A plugin that interacts with the project [BF1ClientAPI]"
authors = ["SAFEluren <safeluren@Outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
nonebot-plugin-localstore = "^0.6.0"
nonebot2 = {extras = ["fastapi"], version = "^2.2.1"}
nonebot-adapter-onebot = "^2.4.1"
zhconv = "^1.4.3"
nonebot-plugin-alconna = "^0.38.2"
nonebot-plugin-apscheduler = "^0.4.0"
toml = "^0.10.2"
nonebot-plugin-access-control-api = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
